backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-dotenv
requests
apscheduler
firecrawl-py
google-genai
pydantic